add strategy

+++ contracts/Module/interfaces/IGovernorModule.sol
@@ -146,7 +146,7 @@ contract GovernorBravoDelegateStorageV1 is GovernorBravoDelegatorStorage {
         /// @notice Whether or not the voter supports the proposal or abstains
         uint8 support;
         /// @notice The number of votes the voter had, which were cast
-        uint96 votes;
+        uint256 votes;
     }
 
     /// @notice Possible states that a proposal may be in
